requirements.txt
dictor
requests
opencv-python==4.2.0.32
pygame
tabulate
pexpect
transforms3d
pandas
py-trees==0.8.3
networkx
numpy
PyYAML
